[feat] relax constrain on contiguous multicast able to use partial chain
