+++ tests/dataflow/aie/experimental/run_all.sh
@@ -6,7 +6,8 @@ PYTHON=python3
 
 FILES=(
     test_gemm.py
-    test_mapping.py
+    test_mapping_basic.py
+    test_mapping_gemm.py
     test_matmul.py
     test_matrix_add.py
     test_pingpong_gemm.py
